updated Description and Title field in DESCRIPTION
Updated Description and Title field according to description in CRAN
email:

Please also make sure that in your package DESCRIPTION file,

* The Title field uses title case and does not end a period.  It should
not repeat the package name, and should be on one line and preferably
no more than 65 chars.

* The Description field really give a comprehensive description of what
the package does, and avoids redundancies such as "This package ..."
or "An R package for ...".